Refactored config for tests to use st2tests.config.
Verified that tests run individually. Also marked places
where config needs to be setup even before importing some
modules. We need to fix this.
