+++ st2actions/tests/test_runner_container_service.py
@@ -10,9 +10,14 @@ import unittest2
 
 from st2actions.constants import LIBS_DIR as ACTION_LIBS_DIR
 from st2actions.container.service import RunnerContainerService
+import st2tests.config as tests_config
 
 
 class RunnerContainerServiceTest(unittest2.TestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        tests_config.parse_args()
 
     def test_get_entry_point_absolute_path(self):
         service = RunnerContainerService()
