+++ st2auth/tests/base.py
@@ -1,8 +1,8 @@
 from webtest import TestApp
 
-import tests.config
 from st2tests import DbTestCase
 from st2auth import app
+import st2tests.config as tests_config
 
 
 class FunctionalTest(DbTestCase):
@@ -10,7 +10,7 @@ class FunctionalTest(DbTestCase):
     @classmethod
     def setUpClass(cls):
         super(FunctionalTest, cls).setUpClass()
-        tests.config.parse_args()
+        tests_config.parse_args()
         cls.app = TestApp(app.setup_app())
 
     @classmethod
